+++ src/main/scala/io/iohk/ethereum/App.scala
@@ -7,7 +7,9 @@ import akka.actor.ActorSystem
 import akka.agent._
 import akka.util.ByteString
 import io.iohk.ethereum.crypto._
-import io.iohk.ethereum.network.{PeerManagerActor, ServerActor}
+import io.iohk.ethereum.db.dataSource.EphemDataSource
+import io.iohk.ethereum.db.storage._
+import io.iohk.ethereum.network.{FastSyncActor, PeerManagerActor, ServerActor}
 import io.iohk.ethereum.utils.{BlockchainStatus, Config, NodeStatus, ServerStatus}
 import org.spongycastle.util.encoders.Hex
 
@@ -38,7 +40,15 @@ object App {
       peerManager ! PeerManagerActor.ConnectToPeer(node)
       //TODO change to CLI command?
       Thread.sleep(2 * 1000)
-      peerManager ! PeerManagerActor.StartFastDownload(node, ByteString(Hex.decode("12d7b70f28b819867087c37e4190b727dfbb1f7f34e2687e5ad126a31cf051be")))
+      val storage = FastSyncActor.Storage(
+        new BlockHeadersStorage(EphemDataSource()),
+        new BlockBodiesStorage(EphemDataSource()),
+        new ReceiptStorage(EphemDataSource()),
+        new MptNodeStorage(EphemDataSource()),
+        new EvmCodeStorage(EphemDataSource())
+      )
+      peerManager ! PeerManagerActor.StartFastDownload(node,
+        ByteString(Hex.decode("12d7b70f28b819867087c37e4190b727dfbb1f7f34e2687e5ad126a31cf051be")), storage)
     }
   }
